Add files via upload
The main motive of MyGuruG.com is to guide the students of who are financially not up the par and providing them a platform where they can get chance to learn. The needy students can have free access to the site. Some of the students face problem in understanding English, so this site provide a wonderful platform to learn the concepts. Language doesn't matter at early stages, but concepts matters without which a student may have face a lot of difficulties. They can learn communication & speaking skill later. But once their leaning period passed away it will become a huge problem for them.