Add files via upload
The motive of the MyGururG.com is to provide a platform for the needy students who are financially poor. This portal provide Videos & PDFs. Some of the students who feels that they have problem in understanding English, so this portal provide a free access to them.

The goal behind this is "Education for free". Hope it will achieve a great success.